Rename p2p.ts to connection.ts
